Revert "refs"
This reverts commit c64a153f67cd26daa58c3bf5f911da05d41cfb5d.

+++ beacon-chain/state/state_trie.go
@@ -255,12 +255,12 @@ func (b *BeaconState) FieldReferencesCount() map[string]uint64 {
 	b.lock.RLock()
 	defer b.lock.RUnlock()
 	for i, f := range b.sharedFieldReferences {
-		refMap[i.String()] = f.Refs()
+		refMap[i.String()] = uint64(f.Refs())
 	}
 	for i, f := range b.stateFieldLeaves {
 		f.lock.RLock()
 		if len(f.fieldLayers) != 0 {
-			refMap[i.String()+"_trie"] = f.Refs()
+			refMap[i.String()+"_trie"] = uint64(f.Refs())
 		}
 		f.lock.RUnlock()
 	}
